Removed hard coded API settings.

+++ src/ca/mphost/importer/CreatePerson.java
@@ -3,6 +3,9 @@
  */
 package ca.mphost.importer;
 
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
+
 import ca.mphost.importer.nationbuilder.Person;
 
 /**
@@ -17,31 +20,43 @@ public class CreatePerson {
 	 * @param args
 	 */
 	public static void main(String[] args) throws Exception {
-		if( args.length < 3 ) {
-			System.out.println("Usage: ca.mphost.importer.CreateEvent name sandbox apikey");
-			//System.exit( 1 );
+		
+		if( args.length < 2 ) {
+			System.out.println("Usage: ca.mphost.importer.CreatePerson sandbox apikey");
+			System.exit( 1 );
 		}
-		String sandbox = "bmulvihillsandbox";
-		String site = "testeda";
-		String apikey = "eb243fcc528ad9720ea0ba451c61500f1cbd85c074c463dbe5a164576d95de17";
+		String sandbox = args[0];
+		String apikey = args[1];
 		
 		NationBuilderAPI api = new NationBuilderAPI();
-		api.init(sandbox, site, apikey);
+		api.init(sandbox, null, apikey);
 		
 		Person person = new Person();
-		person.setFirst_name( "Ken");
-		person.setLast_name( "Ong" );
-		person.setEmail("kenong@test.com");
-		person.setSex( "M" );
+		
+		System.out.println("First name:");
+		BufferedReader reader = new BufferedReader( new InputStreamReader( System.in ) );
+		String firstName = reader.readLine();
+		
+		person.setFirst_name( firstName );
+		System.out.println("Last name");
+		person.setLast_name( reader.readLine() );
+		System.out.println("Email address");
+		person.setEmail( reader.readLine() );
+		System.out.println("Sex [M/F]");
+		person.setSex( reader.readLine() );
 		
 		//Create the person
 		Person ret = api.createPerson(person);
 		//Update the person
-		ret.setEmail( "kenong2@test.com");
+		System.out.println( "New email address");
+		ret.setEmail( reader.readLine() );
 		api.updatePerson(ret);
 		
 		//Cleanup.
-		api.deletePerson(ret);
+		System.out.println( "Cleanup [Y/N]");
+		String confirm = reader.readLine();
+		if( confirm.equals( "Y" ))
+			api.deletePerson(ret);
 
 	}
